show runstate on instruments page
