+++ app/components/ScienceGroup.tsx
@@ -1,0 +1,28 @@
+import { IfcInstrumentStatus } from "@/app/types";
+import InstrumentWallCard from "@/app/components/InstrumentWallCard";
+
+export default function ScienceGroup({
+  name,
+  instruments,
+}: {
+  name: string;
+  instruments: Array<IfcInstrumentStatus>;
+}) {
+  return (
+    <div
+      key={name}
+      className={"flex flex-col justify-center items-start w-full"}
+    >
+      <h2 className="w-full text-left text-gray-600 dark:text-gray-200 font-semibold text-md mt-2 py-2 ">
+        {name}
+      </h2>
+      <div className={"flex flex-wrap gap-1"}>
+        {instruments.map((instrument: IfcInstrumentStatus) => {
+          return (
+            <InstrumentWallCard key={instrument.name} instrument={instrument} />
+          );
+        })}
+      </div>
+    </div>
+  );
+}
