+++ app/components/ShowHideBeamInfo.tsx
