+++ app/wall/page.tsx
@@ -1,152 +1,9 @@
 "use client";
-import { useEffect, useState } from "react";
-import useWebSocket from "react-use-websocket";
-import { instListFromBytes } from "../components/dehex_and_decompress";
-import InstrumentGroup from "./components/InstrumentGroup";
-import ShowHideBeamInfo from "./components/ShowHideBeamInfo";
-import JenkinsJobIframe from "./components/JenkinsJobsIframe";
-import { IfcPVWSMessage, IfcPVWSRequest, targetStation } from "@/app/types";
-import { instListPV, instListSubscription, socketURL } from "@/app/commonVars";
-import {
-  updateInstrumentRunstate,
-  updateInstrumentRunstatePV,
-} from "@/app/wall/utils";
+import ShowHideBeamInfo from "../components/ShowHideBeamInfo";
+import JenkinsJobIframe from "../components/JenkinsJobsIframe";
+import InstrumentsDisplay from "@/app/components/InstrumentsDisplay";
 
 export default function WallDisplay() {
-  const runstatePV = "DAE:RUNSTATE_STR";
-
-  const [data, setData] = useState<Array<targetStation>>([
-    {
-      targetStation: "Target Station 1",
-      instruments: [
-        { name: "ALF" },
-        { name: "CRISP" },
-        { name: "EMMA-A" },
-        { name: "EMU" },
-        { name: "ENGINX" },
-        { name: "GEM" },
-        {
-          name: "HIFI-CRYOMAG",
-        },
-        { name: "HRPD" },
-        { name: "INES" },
-        { name: "IRIS" },
-        { name: "LOQ" },
-        { name: "MAPS" },
-        { name: "MARI" },
-        { name: "MERLIN" },
-        { name: "MUONFE" },
-        { name: "MUSR" },
-        { name: "OSIRIS" },
-        { name: "PEARL" },
-        { name: "POLARIS" },
-        { name: "RIKENFE" },
-        { name: "SANDALS" },
-        { name: "SCIDEMO" },
-        { name: "SURF" },
-        { name: "SXD" },
-        { name: "TOSCA" },
-        { name: "VESUVIO" },
-      ],
-    },
-    {
-      targetStation: "Target Station 2",
-      instruments: [
-        { name: "CHIPIR" },
-        { name: "IMAT" },
-        { name: "INTER" },
-        { name: "LARMOR" },
-        { name: "LET" },
-        { name: "NIMROD" },
-        { name: "OFFSPEC" },
-        { name: "POLREF" },
-        { name: "SANS2D" },
-        { name: "WISH" },
-        { name: "ZOOM" },
-      ],
-    },
-    {
-      targetStation: "Miscellaneous",
-      instruments: [
-        { name: "ARGUS" },
-        { name: "CHRONUS" },
-        {
-          name: "CRYOLAB_R80",
-        },
-        { name: "DCLAB" },
-        { name: "DEMO" },
-        { name: "DETMON" },
-        {
-          name: "ENGINX_SETUP",
-        },
-        { name: "HIFI" },
-        {
-          name: "HRPD_SETUP",
-        },
-        {
-          name: "IBEXGUITEST",
-        },
-        {
-          name: "IRIS_SETUP",
-        },
-        { name: "MOTION" },
-        {
-          name: "PEARL_SETUP",
-        },
-        { name: "SELAB" },
-        { name: "SOFTMAT" },
-        {
-          name: "WISH_SETUP",
-        },
-      ],
-    },
-  ]);
-
-  const {
-    sendJsonMessage,
-    lastJsonMessage,
-  }: {
-    sendJsonMessage: (a: IfcPVWSRequest) => void;
-    lastJsonMessage: IfcPVWSMessage;
-  } = useWebSocket(socketURL, {
-    shouldReconnect: (closeEvent) => true,
-  });
-
-  useEffect(() => {
-    // On page load, subscribe to the instrument list as it's required to get each instrument's PV prefix.
-    sendJsonMessage(instListSubscription);
-  }, [sendJsonMessage]);
-
-  useEffect(() => {
-    // This is a PV update, it could be either the instlist or an instrument's runstate that has changed
-    if (!lastJsonMessage) {
-      return;
-    }
-
-    const updatedPV: IfcPVWSMessage = lastJsonMessage;
-    const updatedPVName: string = updatedPV.pv;
-    const updatedPVbytes: string | null | undefined = updatedPV.b64byt;
-    let updatedPVvalue: string | null | undefined = updatedPV.text;
-
-    if (updatedPVName == instListPV && updatedPVbytes != null) {
-      const instListDict = instListFromBytes(updatedPVbytes);
-      for (const instrument of instListDict) {
-        setData((prev) => {
-          return updateInstrumentRunstatePV(
-            prev,
-            instrument,
-            runstatePV,
-            sendJsonMessage,
-          );
-        });
-      }
-    } else if (updatedPVvalue) {
-      setData((prev) => {
-        return updateInstrumentRunstate(prev, updatedPVName, updatedPVvalue);
-      });
-    }
-  }, [lastJsonMessage, sendJsonMessage]);
-
   return (
     <main
       className={`flex min-h-screen bg-white dark:bg-zinc-800 flex-col items-center justify-between`}
@@ -162,16 +19,7 @@ export default function WallDisplay() {
               <h1 className="w-full text-left text-black dark:text-white font-semibold text-2xl   ">
                 Instrument Status:
               </h1>
-
-              {data.map((targetStation) => {
-                return (
-                  <InstrumentGroup
-                    key={targetStation.targetStation}
-                    groupName={targetStation.targetStation}
-                    data={targetStation.instruments}
-                  />
-                );
-              })}
+              <InstrumentsDisplay />
             </div>
           </div>
           <hr className="h-[2px] rounded my-4 bg-gray-200 border-0 dark:bg-gray-600" />
